# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: NinerLog API
  version: 1.0.0
  description: |
    RESTful API for EASA/FAA compliant pilot logbook management.
    
    ## Features
    - Multi-license support (PPL, SPL, CPL, ATPL, IR)
    - EASA and FAA regulatory compliance
    - Flight logging with detailed time tracking
    - Currency calculations
    - Statistics and hour totals per license
    
    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require JWT authentication.
    Include the token in the Authorization header: `Bearer <token>`
    
  contact:
    name: Frederic Jaeckel
    email: frederic@ninerlog.app
  license:
    name: TBD
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://app.ninerlog.com/api/v1
    description: Production server
...
paths:
...
  /admin/stats:
    get:
      tags:
        - Admin
      summary: System-wide aggregate statistics
      description: |
        Returns aggregate platform metrics. No individual user data exposed.
        Requires admin privileges.
      operationId: getAdminStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/maintenance/cleanup-tokens:
    post:
      tags:
        - Admin
      summary: Clean up expired tokens
      description: Delete expired refresh tokens and password reset tokens. Requires admin privileges.
      operationId: cleanupTokens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cleanup results
          content:
            application/json:
              schema:
                type: object
                properties:
                  refreshTokensDeleted:
                    type: integer
                  resetTokensDeleted:
                    type: integer
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/maintenance/smtp-test:
    post:
      tags:
        - Admin
      summary: Send SMTP test email
      description: Send a test email to the admin's email address to verify SMTP configuration. Requires admin privileges.
      operationId: smtpTest
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Test email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/maintenance/trigger-notifications:
    post:
      tags:
        - Admin
      summary: Trigger notification check
      description: |
        Immediately run the notification checker for all users, bypassing the
        check-hour gating. Used for testing and manual maintenance.
        Requires admin privileges.
      operationId: triggerNotifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification check triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/config:
    get:
      tags:
        - Admin
      summary: Runtime configuration (non-secret)
      description: |
        Returns non-secret runtime configuration for system diagnostics.
        Never exposes secrets (JWT keys, SMTP passwords, database credentials).
        Requires admin privileges.
      operationId: getAdminConfig
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Runtime configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
...
  schemas:
...
    AdminStats:
      type: object
      required:
        - totalUsers
        - totalFlights
        - totalAircraft
        - totalCredentials
        - totalImports
        - flightsThisMonth
        - newUsersThisWeek
        - lockedAccounts
        - disabledAccounts
        - cloudBackupDestinations
      properties:
        totalUsers:
          type: integer
        totalFlights:
          type: integer
        totalAircraft:
          type: integer
        totalCredentials:
          type: integer
        totalImports:
          type: integer
        flightsThisMonth:
          type: integer
        newUsersThisWeek:
          type: integer
        lockedAccounts:
          type: integer
        disabledAccounts:
          type: integer
        cloudBackupDestinations:
          type: object
          description: Counts of user-configured cloud backup destinations.
          required:
            - total
            - byProvider
          properties:
            total:
              type: integer
              description: Total number of cloud backup destinations configured across all users.
            byProvider:
              type: object
              additionalProperties:
                type: integer
              description: 'Number of destinations grouped by provider name (e.g. {"s3": 3, "sftp": 1}). Providers with zero destinations are omitted.'
              example:
                s3: 3
                sftp: 1
...
    AdminConfig:
      type: object
      required:
        - goVersion
        - serverUptime
        - migrationVersion
        - airportDatabaseSize
        - corsOrigins
        - rateLimitAuth
        - rateLimitAdmin
        - smtpConfigured
        - adminEmailConfigured
        - cloudBackupsConfigured
        - cloudBackupProviders
      properties:
        goVersion:
          type: string
          example: go1.23.0
        serverUptime:
          type: string
          description: Human-readable uptime
          example: 2d 5h 30m
        migrationVersion:
          type: integer
          description: Current database migration version
          example: 28
        airportDatabaseSize:
          type: integer
          description: Number of airports loaded
          example: 29331
        corsOrigins:
          type: array
          items:
            type: string
          description: Configured CORS allowed origins
        rateLimitAuth:
          type: string
          description: Auth endpoint rate limit
          example: 10 req/min
        rateLimitAdmin:
          type: string
          description: Admin endpoint rate limit
          example: 30 req/min
        smtpConfigured:
          type: boolean
          description: Whether SMTP is configured
        adminEmailConfigured:
          type: boolean
          description: Whether ADMIN_EMAIL is set
        cloudBackupsConfigured:
          type: boolean
          description: Whether cloud backups are enabled (BACKUP_CREDENTIALS_KEY is set)
        authMode:
          type: string
          enum: [local, oidc]
          description: Active authentication mode (oidc when OIDC_ISSUER is set)
        oidcIssuer:
          type: string
          description: >-
            Configured OIDC issuer URL. Present only in oidc mode; the client
            ID and secret are never exposed.
        cloudBackupProviders:
          type: array
          items:
            type: string
          description: Names of registered cloud backup providers (e.g. s3, sftp, webdav). Empty when cloud backups are disabled.
          example: [s3, sftp, webdav]
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Validation failed
          description: Error message
        code:
          type: string
          example: email_not_verified
          description: |
            Optional machine-readable error code, used by clients to distinguish
            between conditions that share an HTTP status (e.g. `email_not_verified`
            vs `account_disabled` for 403 on /auth/login).
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: totalTime
              message:
                type: string
                example: Must be greater than 0
          description: Detailed validation errors
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    Forbidden:
      description: Forbidden - admin access required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Admin access required
...